third_party/WebKit/LayoutTests/http/tests/origin_trials/webexposed/weblocksapi-origin-trial-interfaces.html

<!DOCTYPE html>
<meta charset="utf-8">
<!-- Generate token with the command:
generate_token.py http://127.0.0.1:8000 WebLocksAPI --expire-timestamp=2000000000
-- -->
<meta http-equiv="origin-trial" content="Aq40kr/ZTqxmfeh35cvBQcwBrmiL7pSDR6PrUZaVC7xxGe3ff4fECD/TdP+w+Ic9cXZ1ek6N4kg6oR876PQd/QoAAABTeyJvcmlnaW4iOiAiaHR0cDovLzEyNy4wLjAuMTo4MDAwIiwgImZlYXR1cmUiOiAiV2ViTG9ja3NBUEkiLCAiZXhwaXJ5IjogMjAwMDAwMDAwMH0=" />
<title>Web Locks API - interfaces exposed by origin trial</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/origin-trials-helper.js"></script>
<script>
test(t => {
  OriginTrialsHelper.check_properties(this,
      {'LockManager': ['request', 'query'],
       'Lock': ['name', 'mode'],
       });
  assert_true('locks' in self.navigator, 'locks property exists on navigator');
}, 'Web Locks API interfaces and properties in Origin-Trial enabled document.');

</script>
